Infrastructure Omnibus PR  (#4)
Ticket
https://wicmtdp.atlassian.net/browse/{TICKET_ID}

Changes
🚚 Consolidates the infrastructure terraform files added in Create Terraform resources for eligibility screener wic-mt-demo-project-eligibility-screener#6 and the file structure set up in the mock api repo
Adds ECS/ECR services to both applications
Adds related resources for services (e.g. IAM roles and security groups)
Context for reviewers
As the project grows, the decision was made to move all of the infrastructure-related items to one place. The documentation currently related to this structure should be updated in a future ticket. Right now, the way things are can be split into two distinct modules.
eligibility_screener holds all of the infrastructure that was in the eligibility screener repo. This also currently hold all of the code relating to the ECS service for the front-end
mock_api hold all of the infrastructure for the mock api. A lot of this is new, and not in the mock api repo